# docs/openAPI.yaml
openapi: 3.0.3
info:
  title: NFC Crypto Payment API
  version: "1.0.0"
  description: >
    API for the NFC crypto payment system. This API provides:
      - User registration with JWT issuance.
      - User login with JWT.
      - Submission of new wallet information (wallet name, coin-address mapping, and ERC-20 tokens).
      - Retrieval of all wallet IDs for a user with coin/token addresses.
      - Retrieval of wallet balances (including tokens).
      - Submission of a signed transaction (raw hex string).
      - Retrieval of transaction status using store ID and transaction ID.
      - Terminal authentication and transaction management (creating and closing transactions).
servers:
  - url: https://api.yourdomain.com
    description: Production server

paths:
  /auth/register:
    post:
      summary: "User Registration"
      description: "Register a new user with first name, last name, email, and password. Returns a user ID and a JWT."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        '200':
          description: "User successfully registered"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
        '400':
          description: "Registration error"

  /auth/login:
    post:
      summary: "User Login"
      description: "Authenticate user with email and password. Returns a JWT."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        '200':
          description: "Login successful"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        '401':
          description: "Authentication error"

  /wallets:
    post:
      summary: "Add New Wallet"
      description: "Submit new wallet information. The request includes the user ID, wallet name, a mapping of coin identifiers to wallet addresses, and supported ERC-20 tokens."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddWalletRequest"
      security:
        - BearerAuth: [ ]  # JWT authentication
      responses:
        '200':
          description: "Wallet successfully added"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddWalletResponse"
        '400':
          description: "Error adding wallet"

  /wallets/list:
    post:
      summary: "Retrieve All Wallets"
      description: "Get all wallet IDs and names associated with a user, along with coin and token addresses."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletListRequest"
      security:
        - BearerAuth: [ ]  # JWT authentication
      responses:
        '200':
          description: "List of wallets retrieved"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletListResponse"
        '400':
          description: "Error retrieving wallet list"

  /wallets/balance:
    post:
      summary: "Get Wallet Balance"
      description: "Retrieve the current balances for all coins (including ERC-20 tokens) in a specific wallet."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletBalanceRequest"
      security:
        - BearerAuth: [ ]  # JWT authentication
      responses:
        '200':
          description: "Wallet balance retrieved"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalanceResponse"
        '400':
          description: "Error retrieving balance"

  /transactions:
    post:
      summary: "Submit Signed Transaction"
      description: "Submit a signed transaction (raw hex string). The server validates the transaction and returns a transaction ID and status."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendTransactionRequest"
      security:
        - BearerAuth: [ ]  # JWT authentication
      responses:
        '200':
          description: "Transaction accepted for processing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendTransactionResponse"
        '400':
          description: "Error submitting transaction"

  /transactions/status:
    post:
      summary: "Get Transaction Status"
      description: "Retrieve the status of a transaction by providing both store ID and transaction ID."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionStatusRequest"
      security:
        - BearerAuth: [ ]  # JWT authentication
      responses:
        '200':
          description: "Transaction status retrieved"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionStatusResponse"
        '404':
          description: "Transaction not found"

  /terminals/authenticate:
    post:
      summary: "Terminal Authentication"
      description: "Authenticate a terminal to receive a JWT and store ID."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminalAuthRequest"
      responses:
        '200':
          description: "Terminal authenticated successfully"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalAuthResponse"
        '401':
          description: "Authentication error"

  /terminals/transactions:
    post:
      summary: "Create Transaction by Terminal"
      description: "A terminal creates a transaction with transaction ID and timestamp."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminalTransactionRequest"
      security:
        - BearerAuth: [ ]  # JWT authentication for terminal
      responses:
        '200':
          description: "Transaction created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalTransactionResponse"
        '400':
          description: "Error creating transaction"

  /terminals/transactions/close:
    post:
      summary: "Close Transaction by Terminal"
      description: "A terminal closes a transaction by providing the transaction ID."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TerminalCloseTransactionRequest"
      security:
        - BearerAuth: [ ]  # JWT authentication for terminal
      responses:
        '200':
          description: "Transaction closed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TerminalCloseTransactionResponse"
        '400':
          description: "Error closing transaction"

components:
  schemas:
    RegisterRequest:
      type: object
      required:
        - fname
        - lname
        - email
        - password
      properties:
        fname:
          type: string
          example: "John"
        lname:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        password:
          type: string
          format: password
          example: "StrongPassword123!"
    RegisterResponse:
      type: object
      properties:
        user_id:
          type: string
          example: "user123"
        token:
          type: string
          description: "JWT for authentication"
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        password:
          type: string
          format: password
          example: "StrongPassword123!"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user_id:
          type: string
          example: "user123"
    AddWalletRequest:
      type: object
      required:
        - user_id
        - wallet_name
        - coins
      properties:
        user_id:
          type: string
          example: "user123"
        wallet_name:
          type: string
          example: "Main Wallet"
        coins:
          type: object
          description: "Mapping of coin identifier to wallet address"
          additionalProperties:
            type: string
          example:
            ETH: "0xabc123..."
            BTC: "1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa"
        tokens:
          type: object
          description: "Mapping of ERC-20 token identifiers to wallet addresses"
          additionalProperties:
            type: string
          example:
            DAI: "0x1234567890abcdef1234567890abcdef12345678"
    AddWalletResponse:
      type: object
      properties:
        wallet_id:
          type: string
          example: "wallet456"
        message:
          type: string
          example: "Wallet successfully added"
    WalletListRequest:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
          example: "user123"
    WalletListResponse:
      type: object
      properties:
        wallets:
          type: array
          items:
            type: object
            properties:
              wallet_id:
                type: string
                example: "wallet456"
              wallet_name:
                type: string
                example: "Main Wallet"
              coins:
                type: object
                description: "Coins and their addresses in the wallet"
                additionalProperties:
                  type: string
                example:
                  ETH: "0xabc123..."
                  BTC: "1A1zP1eP5QGefi2DMPTfTL5SLmv7DivfNa"
              tokens:
                type: object
                description: "ERC-20 tokens and their addresses"
                additionalProperties:
                  type: string
                example:
                  DAI: "0x1234567890abcdef1234567890abcdef12345678"
    WalletBalanceRequest:
      type: object
      required:
        - wallet_id
      properties:
        wallet_id:
          type: string
          example: "wallet456"
    WalletBalanceResponse:
      type: object
      properties:
        wallet_id:
          type: string
          example: "wallet456"
        balances:
          type: object
          description: "Mapping of coin identifier to balance"
          additionalProperties:
            type: string
          example:
            ETH: "1.2345"
            BTC: "0.0567"
            DAI: "500.00"
    SendTransactionRequest:
      type: object
      required:
        - user_id
        - transaction
      properties:
        user_id:
          type: string
          example: "user123"
        transaction:
          type: string
          description: "Signed transaction in raw hex string format"
          example: "0xf86c80850ba43b740083015f9094..."
    SendTransactionResponse:
      type: object
      properties:
        transaction_id:
          type: string
          example: "tx789"
        status:
          type: string
          description: "Validation status of the transaction (accepted, pending, failed)"
          example: "accepted"
        message:
          type: string
          example: "Transaction accepted for processing"
    TransactionStatusRequest:
      type: object
      required:
        - transaction_id
      properties:
        transaction_id:
          type: string
          description: "Unique transaction ID"
          example: "tx789"
    TransactionStatusResponse:
      type: object
      properties:
        transaction_id:
          type: string
          example: "tx789"
        status:
          type: string
          description: "Transaction status (pending, confirmed, failed)"
          example: "confirmed"
        details:
          type: object
          description: "Additional details about the transaction"
          additionalProperties: true
          example:
            blockNumber: 12345678
            timestamp: "2025-02-04T12:45:00Z"

    # Terminal API
    TerminalAuthRequest:
      type: object
      required:
        - terminal_id
        - store_secret
      properties:
        terminal_id:
          type: string
          example: "terminal001"
        store_secret:
          type: string
          example: "supersecretkey"
    TerminalAuthResponse:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        store_id:
          type: string
          example: "store123"

    TerminalTransactionRequest:
      type: object
      required:
        - store_id
        - transaction_id
        - timestamp
      properties:
        store_id:
          type: string
          example: "store123"
        transaction_id:
          type: string
          example: "tx001"
        timestamp:
          type: string
          format: date-time
          example: "2025-02-04T12:45:00Z"

    TerminalTransactionResponse:
      type: object
      properties:
        message:
          type: string
          example: "Transaction created successfully"
        transaction_id:
          type: string
          example: "tx001"

    TerminalCloseTransactionRequest:
      type: object
      required:
        - store_id
        - transaction_id
      properties:
        store_id:
          type: string
          example: "store123"
        transaction_id:
          type: string
          example: "tx001"

    TerminalCloseTransactionResponse:
      type: object
      properties:
        message:
          type: string
          example: "Transaction closed successfully"